+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## Version 9.1.6 (??)
+* Fix: Various fixes and cleanup on Linux Firewire and DVB.
+* Fix: Added support for all 4 tuners on the Hauppauge WinTV-quadHD tuner in Windows.
+* New: Add Schedules Direct lineup by ID.
+* Change: Removed ZZZ from Schedules Direct Regions because it doesn't do anything.
+
 ## Version 9.1.5 (2017-06-19)
 * Fix: Carny throws a null pointer exception if a show has a null title.